--- src/components/sections/Hero/index.js
@@ -1,2 +1,0 @@
-// src/components/sections/Hero/index.js
-export { default } from './Hero';
