--- src/components/sections/index.js
@@ -1,6 +1,0 @@
-// src/components/sections/index.js
-export { default as Hero } from './Hero/Hero';
-export { default as Skills } from './Skills';
-export { default as Experience } from './Experience/Experience';
-export { default as Projects } from './Projects';
-export { default as Education } from './Education';
