+++ src/features/certifications/components/CertCard.desktop.jsx
