--- src/mobileComponents/common/SectionWrapper/types.js
@@ -1,15 +1,0 @@
-// src/components/common/SectionWrapper/types.js
-/**
- * @typedef {Object} SectionWrapperProps
- * @property {React.ReactNode} children - The content to be wrapped
- * @property {string} [className] - Additional CSS classes
- * @property {'glass' | 'transparent'} [variant='transparent'] - The visual variant
- * @property {boolean} [animate=true] - Whether to animate on scroll
- */
-
-/**
- * @typedef {Object} VariantStyles
- * @property {string} background - Background CSS class
- * @property {string} border - Border CSS class
- * @property {string} backdrop - Backdrop filter CSS class
- */
