--- src/mobileComponents/sections/Goals/utils/styles.js
@@ -1,7 +1,0 @@
-// src/components/sections/Goals/utils/styles.js
-export const gradientStyles = {
-    learning: 'from-emerald-500 to-cyan-500',
-    Certifications: 'from-amber-500 to-orange-500',
-    projects: 'from-violet-500 to-purple-500',
-    career: 'from-pink-500 to-rose-500'
-};
