--- src/mobileComponents/sections/index.js
@@ -1,6 +1,0 @@
-// src/components/sections/index.js
-export { default as MobileHero } from './Hero/Hero';
-export { default as MobileSkills } from './Skills/Skills';
-export { default as MobileExperience } from './Experience/MobileExperience';
-export { default as MobileProjects } from './Projects/Projects';
-export { default as MobileEducation } from './Education/Education';
